backend/requirements.txt
flask
flask_sqlalchemy
python-dotenv
flask_jwt_extended
flask_cors
requests
boto3
redis
payos
psycopg2-binary
sqlalchemy
gunicorn
